// src/interface/interface.ts
export interface ticketsType {
  origin: string | number;
  origin_name: string;
  destination: string;
  destination_name: string;
  departure_date: string;
  departure_time: string;
  arrival_date: string;
  arrival_time: string;
  carrier: string;
  stops: number;
  price: number;
  cost?: number;
  val?: string;
}